+++ pkg/metadata/metadata.go
@@ -23,6 +23,7 @@ type BackupMetadata struct {
 	CompressedSize          uint64            `json:"compressed_size,omitempty"`
 	Databases               []DatabasesMeta   `json:"databases,omitempty"`
 	Tables                  []TableTitle      `json:"tables"`
+	Functions               []FunctionsMeta   `json:"functions"`
 	DataFormat              string            `json:"data_format"`
 	RequiredBackup          string            `json:"required_backup,omitempty"`
 }
@@ -31,6 +32,11 @@ type DatabasesMeta struct {
 	Name   string `json:"name"`
 	Engine string `json:"engine"`
 	Query  string `json:"query"`
+}
+
+type FunctionsMeta struct {
+	Name        string `json:"name"`
+	CreateQuery string `json:"create_query"`
 }
 
 type TableMetadata struct {